test: sdk-5057 isolate browser validation coverage
